# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backbon API DOCUMENTATION",
        "description": "This is the Project Progres Information System API documentation.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/identifications": {
            "get": {
                "tags": [
                    "identifications"
                ],
                "summary": "Get a list of identificationss",
                "operationId": "ce290affc5d7bb54d4b1ba46633b060d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "headers": {
                            "Cache-Control": {
                                "description": "Cache control header",
                                "schema": {
                                    "type": "string",
                                    "example": "no-cache, private"
                                }
                            },
                            "Content-Type": {
                                "description": "Content type header",
                                "schema": {
                                    "type": "string",
                                    "example": "application/json; charset=UTF-8"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "identification_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "identification_name": {
                                                        "type": "string",
                                                        "example": "ROLE NATIONAL"
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Mohammed"
                                                    },
                                                    "middle_name": {
                                                        "type": "string",
                                                        "example": "Abdalla"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Bakar"
                                                    },
                                                    "created_by": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-28 11:30:25"
                                                    },
                                                    "deleted_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-28 11:30:25"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-28 11:30:25"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "identifications"
                ],
                "summary": "Store a new identifications",
                "operationId": "22841f0f1f7ad8f5ebed32f8c92bd9b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "identification_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "headers": {
                            "Cache-Control": {
                                "description": "Cache control header",
                                "schema": {
                                    "type": "string",
                                    "example": "no-cache, private"
                                }
                            },
                            "Content-Type": {
                                "description": "Content type header",
                                "schema": {
                                    "type": "string",
                                    "example": "application/json; charset=UTF-8"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "statusCode": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}